Adding Kafka Integration

+++ src/main/java/com/arianthox/predictor/client/BWMatcherClient.java
@@ -1,8 +1,6 @@
 package com.arianthox.predictor.client;
 
-import io.swagger.annotations.ApiOperation;
 import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -16,8 +14,8 @@ import java.util.List;
 @RequestMapping("/api/matcher")
 public interface BWMatcherClient {
     @RequestMapping(value="/add/{key}", method = RequestMethod.PUT, produces="application/json")
-    ResponseEntity addKey(@RequestBody List<short[]> samples, @PathVariable String key);
+    ResponseEntity addKey(@RequestBody List<int[]> samples, @PathVariable String key);
 
     @RequestMapping(value="/match", method = RequestMethod.POST, produces="application/json")
-    HashMap<String, Double> matchKey(@RequestBody short[] signal);
+    HashMap<String, Double> matchKey(@RequestBody int[] signal);
 }
